// src/interfaces/CadastroClienteInterface.ts
export interface CadastroClienteInterface {
    id: number;
    nome: string;
    celular: string;
    email: string;
    cpf: number;
    dataNascimento: string;
    cidade: string;
    estado: string;
    pais: string;
    rua: string;
    numero: string;
    bairro: string,
    cep: string;
    complemento: string;
    senha: string;
}